+++ RNG/RNG.cs
@@ -37,5 +37,6 @@ namespace Orias_RNG
         private long GeneratedNumber;       // The results of the Random Number Generation
         private long LongRand;              // The long number that is randomly generated
         private bool Positive_Only = false; // Flag indicating the results contain only positive values
+        private bool ValidNumber = false;   // Flag to indicate if the results are valid
     } // public class RNG
 } // Orias_RNG