+++ RNG_Test/Tests/Integer_Testing/Signed_Integer/Signed_Integer_Values_Outside_Range.cs
@@ -12,27 +12,24 @@ using Con = System.Console;
 
 namespace RNG_Test
 {
-	partial class BYTE_Testing
+	partial class Integer_Testing
 	{
-		private void Unsigned_BYTE_Negative_Values()
+		private void Signed_Integer_Values_Outside_Range()
 		/*
 		===============================================================================================
 		PURPOSE:
-		Test the Random Number Generator for an Unsigned BYTE when the parameters that are provided
-		are negative.
+		Test the Random Number Generator for and Signed Integer when the parameters that are provided
+		are outside the range of an Unsigned Integer.
 		-----------------------------------------------------------------------------------------------
 		OUTPUT:
-		An error message if the values are outside the range; otherwise a randomly generated Unsigned
-		BYTE value.
+		An error message if the values are outside the range; otherwise a randomly generated Signed
+		Integer value.
 		-----------------------------------------------------------------------------------------------
 		NOTES:
 		- When a value is passed to the RNG, the appropriate overloaded function will be called.
-		  Because of this, a value outside of an Unsigned BYTE becomes a larger unit depending on the
+		  Because of this, a value outside of an Signed Integer becomes a larger unit depending on the
 		  value.  To ensure the RESULTS are not outside the range, the calling process needs to
 		  validate the values PRIOR to calling.
-
-		- Passing a negative value to a Unsigned 8-bit integer causes it to cycle.
-		  For example -1 becomes 255.
 		===============================================================================================
 		*/
 		{
@@ -40,34 +37,43 @@ namespace RNG_Test
 			// Variables - Standard
 			//=============
 			RNG MyRandomNumber = new RNG();
-			byte MinValue = byte.MinValue;
-			byte MaxValue = byte.MaxValue;
+			int MinValue = int.MinValue;
+			int MaxValue = int.MaxValue;
 			bool IsValid = true;  // Assume the value is valid
 
 			//=============
 			// Variables - outside the scope of BYTE
 			//=============
-			int Bad_MinValue = byte.MinValue - 1;
-			int Bad_MaxValue = (byte.MaxValue + 1) * -1;
+			long Bad_MinValue = int.MinValue;
+			long Bad_MaxValue = int.MaxValue;
+
+			//=============
+			// Setup Environment
+			//=============
+			Bad_MinValue -= 1;
+			Bad_MaxValue += 1;
 
 			//=============
 			// Body
 			//=============
 			Con.WriteLine();
-			Con.WriteLine(@"Pass negative values");
+			Con.WriteLine(@"Pass values outside the range");
 
-			if (Bad_MinValue < 0)
+			// Validate the minimum value
+			if (Bad_MinValue < int.MinValue)
 			{
-				Con.WriteLine(@"Intended value ({0}) cannot be converted to UNSIGNED BYTE.", Bad_MinValue);
+				Con.WriteLine(@"Intended value ({0}) cannot be converted to SIGNED Integer.", Bad_MinValue);
 				IsValid = false;
 			}
 
-			if (Bad_MaxValue < 0)
+			// Validate the maximum value
+			if (Bad_MaxValue > sbyte.MaxValue)
 			{
-				Con.WriteLine(@"Intended value ({0}) cannot be converted to UNSIGNED BYTE.", Bad_MaxValue);
+				Con.WriteLine(@"Intended value ({0}) cannot be converted to SIGNED Integer.", Bad_MaxValue);
 				IsValid = false;
 			}
 
+			// Perform the call if we are still valid
 			if (IsValid)
 			{
 				Con.Write("{0}", MyRandomNumber.Generate(Bad_MinValue, MaxValue) + "\t");
@@ -76,6 +82,6 @@ namespace RNG_Test
 				Con.Write("{0}", MyRandomNumber.Generate(MinValue, Bad_MaxValue) + "\t");
 				Con.WriteLine("{0}", MyRandomNumber.Generate(Bad_MinValue, Bad_MaxValue));
 			}
-		} // private void Unsigned_BYTE_Negative_Values
-	} // partial class BYTE_Testing
+		} // private void Signed_Integer_Values_Outside_Range
+	} // partial class Integer_Testing
 } // namespace RNG_Test